update client: refactor client and remove deprecated files
