+++ src/tensorlake/documentai/__init__.py
@@ -1,18 +1,10 @@
 from tensorlake.documentai.common import Document, JobResult
-from tensorlake.documentai.file import Files
-from tensorlake.documentai.jobs import Jobs
-from tensorlake.documentai.parser import (
-    DocumentParser,
-    ExtractionOptions,
-    ParsingOptions,
-)
+from tensorlake.documentai.client import DocumentAI, ParsingOptions, ExtractionOptions
 
 __all__ = [
-    "Files",
-    "DocumentParser",
+    "DocumentAI",
     "ParsingOptions",
     "ExtractionOptions",
-    "Jobs",
     "Document",
     "JobResult",
 ]
